// ts/Dashboard/Component/TextOptions.d.ts
/* *
 *
 *  (c) 2009 - 2023 Highsoft AS
 *
 *  License: www.highcharts.com/license
 *
 *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
 *
 *  Authors:
 *  - Sebastian Bochan
 *  - Wojciech Chmiel
 *  - Gøran Slettemark
 *  - Sophie Bremer
 *
 * */

import type CSSObject from '../../Core/Renderer/CSSObject';
declare interface TextOptions {
    style?: CSSObject;
    text?: string;
    className?: string;
}

export default TextOptions;